fixing code

+++ backend/controllers/websocket.go
@@ -89,8 +89,10 @@ func BroadcastPrivateMessage(msg utils.Message, host string) {
 	errMsg := utils.Err{}
 
 	if ok, err := models.FriendsChecker(msg.Sender_id, msg.Reciever_id); err != nil || !ok {
-		if err := os.Remove("." + msg.Filename); err != nil {
-			fmt.Println(err)
+		if msg.Filename != "" {
+			if err := os.Remove("." + msg.Filename); err != nil {
+				fmt.Println("removin err: ", err)
+			}
 		}
 		errMsg.Error = "you need to follow the receiver first"
 		Broadcast(msg.Sender_id, errMsg)
